Move strategies to another module
