+++ project/blackjack_game/src/player.py
@@ -1,179 +1,19 @@
 from .deck import Deck
 from .card import Card
 from typing import List, TypedDict
+from .strategy import Strategy
 
-# strategies
-class Strategy:
+
+class HistoryDict(TypedDict):
     """
-    Base class for different strategies that a bot can use in the game.
+    Typed dictionary for storing the betting history of a bot player in Blackjack.
 
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck of cards that the bot draws from.
-    """
-
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Abstract method that should be overridden in subclasses to implement the strategy.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        :raises NotImplementedError: If this method is not overridden in a subclass.
-        """
-        raise NotImplementedError("This method should be overridden in subclasses.")
-
-
-class AggressiveStrategy(Strategy):
-    """
-    Aggressive strategy for a bot to keep drawing cards until its score is at least 19.
-
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck from which cards will be drawn.
+    :param bets (List[int]): A list of integers representing the amounts bet by the player in each round.
+    :param results (List[str]): A list of strings representing the outcome of each round ("win", "lose", or "tie").
     """
 
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the aggressive strategy: draws cards until the bot's score is at least 19.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 19:
-            bot.add_card(deck.draw())
-            print(
-                f"{bot.name} ({bot.strategy.__class__.__name__}) draws 🎴 {bot.hand[-1]}."
-            )
-
-
-class ConservativeStrategy(Strategy):
-    """
-    Conservative strategy for a bot to keep drawing cards until its score is at least 15.
-
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck from which cards will be drawn.
-    """
-
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the conservative strategy: draws cards until the bot's score is at least 15.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 15:
-            bot.add_card(deck.draw())
-            print(
-                f"{bot.name} ({bot.strategy.__class__.__name__}) draws 🎴 {bot.hand[-1]}."
-            )
-
-
-class RiskyStrategy(Strategy):
-    """
-    Risky strategy for a bot to keep drawing cards until its score is at least 20.
-
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck from which cards will be drawn.
-    """
-
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the risky strategy: draws cards until the bot's score is at least 20.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 20:
-            bot.add_card(deck.draw())
-            print(
-                f"{bot.name} ({bot.strategy.__class__.__name__}) draws 🎴 {bot.hand[-1]}."
-            )
-
-
-class BasicStrategy(Strategy):
-    """
-    Basic strategy for a bot to keep drawing cards until its score is at least 17.
-
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck from which cards will be drawn.
-    """
-
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the basic strategy: draws cards until the bot's score is at least 17.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 17:
-            bot.add_card(deck.draw())
-            print(
-                f"{bot.name} ({bot.strategy.__class__.__name__}) draws 🎴 {bot.hand[-1]}."
-            )
-
-
-class AdaptiveStrategy(Strategy):
-    """
-    Adaptive strategy that adjusts the bot's behavior based on previous round results.
-
-    :param bot: The bot instance that will make a move.
-    :param deck: The deck from which cards will be drawn.
-    """
-
-    def make_move(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Adjusts the bot's strategy based on the results of the previous round and bet.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        if len(bot.history["results"]) > 0:
-            last_result = bot.history["results"][-1]
-            last_bet = bot.history["bets"][-1]
-
-            if last_result == "lose":
-                print(f"{bot.name} is playing more conservatively after a loss.")
-                self.play_conservative(bot, deck)
-            elif last_result == "win" and last_bet > 50:
-                print(f"{bot.name} is playing aggressively after a big win.")
-                self.play_aggressive(bot, deck)
-            else:
-                print(f"{bot.name} is following a basic strategy.")
-                self.play_basic(bot, deck)
-
-    def play_aggressive(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the aggressive strategy: draws cards until the bot's score is at least 19.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 19:
-            bot.add_card(deck.draw())
-            print(f"{bot.name} (Aggressive) draws 🎴 {bot.hand[-1]}.")
-
-    def play_conservative(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the conservative strategy: draws cards until the bot's score is at least 15.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 15:
-            bot.add_card(deck.draw())
-            print(f"{bot.name} (Conservative) draws 🎴 {bot.hand[-1]}.")
-
-    def play_basic(self, bot: "Bot", deck: Deck) -> None:
-        """
-        Executes the basic strategy: draws cards until the bot's score is at least 17.
-
-        :param bot: The bot instance making the move.
-        :param deck: The deck from which cards will be drawn.
-        """
-        while bot.calculate_score() < 17:
-            bot.add_card(deck.draw())
-            print(f"{bot.name} (Basic) draws 🎴 {bot.hand[-1]}.")
-
-
-# main player class
+    bets: List[int]
+    results: List[str]
 
 
 class Player:
@@ -247,13 +87,11 @@ class Dealer(Player):
         super().__init__("Dealer")
 
 
-class HistoryDict(TypedDict):
-    bets: List[int]
-    results: List[str]
-
-
 class Bot(Player):
     def __init__(self, name: str, strategy: Strategy) -> None:
+        """
+        Initialize a Bot instance.
+        """
         super().__init__(name)
         self.strategy = strategy
         # Define history with clear types for bets and results
@@ -265,6 +103,7 @@ class Bot(Player):
     def add_history(self, bet: int, result: str) -> None:
         """
         Add bet and result information to the history for the current round.
+
         :param bet: The bet amount.
         :param result: The result of the game (win, lose, tie).
         """
